+++ bsn/rollup-finality-provider/cmd/rollup-fpd/main.go
@@ -1,0 +1,53 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"os/signal"
+	"syscall"
+
+	incentivecli "github.com/babylonlabs-io/babylon/v3/x/incentive/client/cli"
+	fpcmd "github.com/babylonlabs-io/finality-provider/finality-provider/cmd"
+	"github.com/babylonlabs-io/finality-provider/finality-provider/cmd/fpd/daemon"
+	fpcfg "github.com/babylonlabs-io/finality-provider/finality-provider/config"
+	"github.com/babylonlabs-io/finality-provider/version"
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/spf13/cobra"
+)
+
+// NewRootCmd creates a new root command for fpd. It is called once in the main function.
+func NewRootCmd() *cobra.Command {
+	rootCmd := &cobra.Command{
+		Use:               "rollup-fpd",
+		Short:             "rollup-fpd - Finality Provider Daemon for rollup BSNs.",
+		Long:              `rollup-fpd is the daemon to create and manage finality providers for rollup BSNs.`,
+		SilenceErrors:     false,
+		PersistentPreRunE: fpcmd.PersistClientCtx(client.Context{}),
+	}
+	rootCmd.PersistentFlags().String(flags.FlagHome, fpcfg.DefaultFpdDir, "The application home directory")
+
+	return rootCmd
+}
+
+func main() {
+	cmd := NewRootCmd()
+	cmd.AddCommand(
+		daemon.CommandInit(), daemon.CommandStart(), daemon.CommandKeys(),
+		daemon.CommandGetDaemonInfo(), daemon.CommandCreateFP(), daemon.CommandLsFP(),
+		daemon.CommandInfoFP(), daemon.CommandAddFinalitySig(), daemon.CommandUnjailFP(),
+		daemon.CommandEditFinalityDescription(), daemon.CommandCommitPubRand(), daemon.CommandRecoverProof(),
+		incentivecli.NewWithdrawRewardCmd(), incentivecli.NewSetWithdrawAddressCmd(),
+		incentivecli.CmdQueryRewardGauges(),
+		version.CommandVersion("fpd"), daemon.CommandUnsafePruneMerkleProof(),
+	)
+
+	ctx, cancel := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
+	defer cancel()
+
+	if err := cmd.ExecuteContext(ctx); err != nil {
+		_, _ = fmt.Fprintf(os.Stderr, "Whoops. There was an error while executing your fpd CLI '%s'", err)
+		os.Exit(1) //nolint:gocritic
+	}
+}